supabase_dw/requirements.txt
Flask
psycopg2-binary
pandas
tqdm
python-dotenv